// client/src/api/generated/models/apiKeyResponseDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Knowted API
 * The Knowted API documentation
 * OpenAPI spec version: 1.0
 */

export interface ApiKeyResponseDto {
  id: string;
  organization_id: string;
  name: string;
  key: string;
  is_active: boolean;
  created_at: string;
  updated_at: string;
}